requirements.txt
biopython==1.74
PyYAML==5.1.2
numpy==1.16.4
lxml==4.4.2
